+++ hoot-core/src/main/cpp/hoot/core/io/OgrReader.cpp
@@ -398,7 +398,7 @@ std::vector<float> OgrReader::_getInputProgressWeights(
   {
     LOG_VART(layers[i]);
     // simply open the file, get the meta feature count value, and close
-    int featuresPerLayer = getFeatureCount(input, layers[i]);
+    long featuresPerLayer = getFeatureCount(input, layers[i]);
     LOG_VART(featuresPerLayer);
     progressWeights.push_back((float)featuresPerLayer);
     // cover the case where no feature count available efficiently; Despite the documentation
